Adds support for adding cmdline options
Adds support for adding cmdline options to the
various kernel cmdlines.  Useful for overriding or
adding options as needed when loading the kernels.

Can be set from the Utilities menu.  Variable added to the end
of the kernel command line to ensure that a long entry doesn't
mess up boot from kernel line character limitations and the
entry getting cut off.

Resolves: https://github.com/netbootxyz/netboot.xyz/issues/652