Adds a hook for putting local-vars.ipxe on tftp
Allows for the user to set local variables to be loaded
into the netboot.xyz menu when loading.  This allows for
setting things like github_user to have custom menu's
auto populate on start up.

Setting an ipxe file with var settings in local-vars.ipxe
of their tftp server would allow those options to be loaded.