Switch to arch variable instead of buildarch
Per iPXE docs, buildarch can be unreliable for architecture
detection, so move our other check earlier and rely
on that for boot.cfg routing.

Have seen issues where tftp booting the image picked up
a 64-bit machine as a 32-bit machine on the bootarch
check and then identifying it with cpuid as a 64-bit machine.

This caused menus to be obscured initially for those machine
types.