Support building images on other architectures
This unconditionally cross compiles all iPXE images. This has the effect
that the build will always produce the correct output for a defined
architecture regardless of host architecture.

syslinux is x86 architecture dependent. It doesn't seem to be used in
the build process, so the packages are removed from the list of required
packages.

Signed-off-by: Felix Kaechele <felix@kaechele.ca>